refactor: modular feature registration
