+++ features/moderation.js
@@ -1,4 +1,5 @@
 const { addBan, removeBan, getBanCollection } = require('../database');
+const { PermissionsBitField } = require('discord.js');
 
 /**
  * Ban a user from a guild and record it in the database.
@@ -63,4 +64,65 @@ async function explainBanQuery(client, message) {
   }
 }
 
-module.exports = { banUser, unbanUser, explainBanQuery };
+function register(client, commands) {
+  commands.set('!ban', '`!ban <@user|userId> [reason]` - Ban a user and record the reason.');
+  commands.set('!unban', '`!unban <userId>` - Remove a ban and unban the user.');
+  commands.set('!banexplain', '`!banexplain` - *Admin only.* Show MongoDB query stats for the ban collection.');
+
+  client.on('messageCreate', async (message) => {
+    try {
+      if (message.author.bot) return;
+      if (!message.content.startsWith('!')) return;
+
+      const args = message.content.trim().split(/\s+/);
+      const command = args.shift().toLowerCase();
+
+      if (command === '!ban') {
+        const user = message.mentions.users.first();
+        const userId = user ? user.id : args[0];
+        if (!userId) return message.reply('Provide a user mention or ID to ban.');
+        const reason = args.slice(1).join(' ') || 'No reason provided';
+        try {
+          await banUser(client, message.guild.id, userId, reason);
+          return message.reply(`Banned ${user ? user.tag : userId}`);
+        } catch (err) {
+          console.error('Ban failed:', err);
+          return message.reply('Failed to ban user.');
+        }
+      }
+
+      if (command === '!unban') {
+        const userId = args[0];
+        if (!userId) return message.reply('Please provide a user ID to unban.');
+        try {
+          await unbanUser(client, message.guild.id, userId);
+          return message.reply(`Unbanned <@${userId}>`);
+        } catch (err) {
+          console.error('Unban failed:', err);
+          return message.reply('Failed to unban user.');
+        }
+      }
+
+      if (command === '!banexplain') {
+        if (!message.member?.permissions.has(PermissionsBitField.Flags.Administrator)) {
+          return message.reply('This command is restricted to administrators.');
+        }
+        try {
+          await explainBanQuery(client, message);
+        } catch (err) {
+          console.error('Explain failed:', err);
+          return message.reply('Failed to retrieve query stats.');
+        }
+      }
+    } catch (err) {
+      console.error('Error handling moderation command:', err);
+      try {
+        await message.reply('An error occurred while processing your command.');
+      } catch (replyErr) {
+        console.error('Failed to send error reply:', replyErr);
+      }
+    }
+  });
+}
+
+module.exports = { register, banUser, unbanUser, explainBanQuery };
